chore: update compose stack and dependency compatibility
Switch the Docker Compose stack from MySQL and Minio to PostgreSQL and rs3,
with the compose-specific configuration updated to use PostgreSQL and the rs3
S3 endpoint.

Adapt code for upgraded resty and GitLab client APIs, including the renamed
response parsing option and int64 pagination and ID fields.

Also refresh agent guidance and the Chinese README automation notes.

Verification:
docker compose -f compose.yml config
ruby -e 'require "yaml"; YAML.load_file("conf/config-compose.yaml"); YAML.load_file("compose.yml"); puts "yaml ok"'
go test -tags "netgo,timetzdata,exclude_graphdriver_btrfs,containers_image_openpgp" ./pkg/config ./pkg/service/token
go test -tags "netgo,timetzdata,exclude_graphdriver_btrfs,containers_image_openpgp" ./pkg/builder ./pkg/distribution/clients ./pkg/background/daemon/webhook
go test -tags "netgo,timetzdata,exclude_graphdriver_btrfs,containers_image_openpgp" ./pkg/background/daemon/coderepo
make build

+++ pkg/service/oauth2/oauth2.go
@@ -170,7 +170,7 @@ func (s *service) Callback(ctx context.Context, provider enums.Provider, code st
 		}
 		userInfo = api.Oauth2UserInfo{
 			Provider:     provider,
-			ID:           strconv.FormatInt(int64(user.ID), 10),
+			ID:           strconv.FormatInt(user.ID, 10),
 			Username:     user.Name,
 			Email:        user.Email,
 			Token:        oauth2Token.AccessToken,
